==> prediction/1-prepare.ipynb <==
# Prepare the selected metapaths for feature computation for all compound-disease pairs

import json
import itertools
import bz2
import configparser

import numpy as np
import pandas
from statsmodels.sandbox.stats.multicomp import multipletests

## Plan

We want to extract the useful metapaths for all chemical disease relations in our training and holdout sets by querying Neo4j. To do so, we will read the training and test sets directly.

## Read training and holdout sets

with open("../../crossval_idx.txt", "r") as fin:
    crossval_idx = int(fin.read().strip())

training = pandas.read_csv("../../../crossval/data/training/training_piece{}.tsv".format(crossval_idx), sep = '\t')
holdout = pandas.read_csv("../../../crossval/data/holdout/holdout_piece{}.tsv".format(crossval_idx), sep = '\t')

training.shape

training.head()

holdout.shape

holdout.head()

holdout["category"].value_counts()

Since the holdout set contains the true labels for whether each relation is a chemical-treats-disease relationship, we will remove this true label but add a flag denoting that the relations are from the holdout set so that we can evaluate them separately later. However they will be combined into the same dataframe for feature extraction from Neo4j in order to avoid drastically modifying the workflow created by Daniel.

## Prepare compound disease pairs tsv

train_pairs = (training[["chemical_id", "chemical_name", "disease_id", "disease_name", "category"]]
    .assign(origin = "train")
)

train_pairs.shape

train_pairs.head()

train_pairs["category"].value_counts()

### Prepare holdout data and remove true labels

holdout.head()

holdout_pairs = (holdout[["chemical_id", "chemical_name", "disease_id", "disease_name", "category"]]
    .assign(origin = "holdout")
    .assign(category = np.nan)
)

holdout_pairs.shape

holdout_pairs.head()

holdout_pairs["category"].isnull().all()

We have successfully sanitized the holdout data pairs to remove the true labels but will be extracting the DWPCs with the training data pairs at the same time.

## Save pairs to file

pair_df = pandas.concat([train_pairs, holdout_pairs])

pair_df["status"] = (pair_df["category"] == "DM").astype(int)

pair_df.shape

pair_df.head()

pair_df["origin"].value_counts()

pair_df.groupby("origin")["category"].value_counts()

with bz2.open('features/compound-disease-pairs.tsv.bz2', 'wt') as write_file:
    pair_df.to_csv(write_file, sep='\t', index=False)

## Select metapaths

auroc_df = pandas.read_table('../all-features/data/feature-performance/auroc.tsv')
reject, pvals_corrected, alphacSidak, alphacBonf = multipletests(auroc_df.pval_delta_auroc, method='fdr_bh')
auroc_df['fdr_delta_auroc'] = pvals_corrected
auroc_df.head(2)

whitelist_df = auroc_df.query(
    "rdwpc_auroc > 0.55"
    " and delta_auroc > 0"
    " and fdr_delta_auroc < 0.05"
    " and pdwpc_primary_auroc > 0.5"
).copy()
whitelist_df['feature'] = 'dwpc_' + whitelist_df['metapath']
whitelist_df.head(2)

metapaths = set(whitelist_df.metapath)
len(metapaths)

## Metaedges in chosen metapaths

m2m_df = pandas.read_table('../all-features/data/metaedge-in-metapath.tsv')
m2m_df = m2m_df.query("metapath in @metapaths")

# Counting multiple occurrences in the same metapath
m2m_df.metaedge.value_counts().reset_index()

# Counting only one metaedge occurrence per metapath
m2m_df.drop_duplicates().metaedge.value_counts().reset_index()

# Number of included metaedges
m2m_df.metaedge.nunique()

with open('../all-features/data/metapaths.json') as read_file:
    metapath_obj = json.load(read_file)
metapath_obj = [x for x in metapath_obj if x['abbreviation'] in metapaths]
with open('features/metapaths.json', 'wt') as write_file:
    metapath_obj = json.dump(metapath_obj, write_file, indent=2, sort_keys=True)
